fix DICOM read with fractional slope
The DICOM reader was multiplying pixel values by rescale_slope, which it
represented as ssize_t. Many DICOM files use double values for
rescale_slope, so for rescale_slope < 1, the reader was multiplying by
zero and producing blank images.

This commit removes all internal handling of rescale_slope and
rescale_intercept, leaving it to the caller to apply these values. It
should be applied to the 7.x branch as well.

This change makes sense within the DICOM standard, since rescale_slope
and rescale_intercept say how to get from the uint16 (or whatever the
file holds) to the value in the units that the file represents, perhaps kBQ/ml.
These are almost always fractional quantities and int-valued pixels
cannot represent them.

Here's a sample file that shows the problem:

http://www.rollthepotato.net/~john/MRIm5.dcm

You're welcome to use this image in a test suite. Before this commit:

$ identify -verbose MRIm5.dcm
  Image: MRIm5.dcm
  Format: DCM (Digital Imaging and Communications in Medicine image)
  Gray:
    min: 32769 (0.500023)
    max: 32769 (0.500023)

After:

$ identify -verbose MRIm5.dcm
  Image: MRIm5.dcm
  Format: DCM (Digital Imaging and Communications in Medicine image)
  Gray:
    min: 0 (0)
    max: 65535 (1)

Forum thread:

https://www.imagemagick.org/discourse-server/viewtopic.php?f=3&t=31887&p=145852