add dcm:window option
This patch adds a new `dcm:window` option which lets you override the
window center and width stored in the DICOM file with a new centerXwidth
pair. You can also use centerX or Xwidth. The window settings are ony
applied if the `dcm:rescale` option is enabled.

This is useful for transforming images from scan units to display units for
visualisation.

Example:

convert -define dcm:rescale -define dcm:window=2500x3000 MRIm5.dcm x.tif

This patch also reverts an earlier change which moved the transform for
signed data into the main path. Window handling assumes a 0 - `max_value`
output range and takes no account of signedness of the data, so
the subtract for signed data is not appropriate. Perhaps it should be,
but we'd need to track `min_value` as well, and that's beyound the scope
of this PR.

Related PR:

https://github.com/ImageMagick/ImageMagick/pull/483

DICOM spec:

https://www.dabsoft.ch/dicom/3/C.11.2.1.2/