simplify boolean expression, found by PVS analyzer
Magick++/lib/Color.cpp	303	warn	V728 An excessive check can be simplified. The '(A && B) || (!A && !B)' expression is equivalent to the 'bool(A) == bool(B)' expression.
Magick++/lib/Image.cpp	314	warn	V728 An excessive check can be simplified. The '(A && !B) || (!A && B)' expression is equivalent to the 'bool(A) != bool(B)' expression.