genesis/terminus: Be more pedantic in the description of `IsMagickCoreInstantiated()'
The  value `MagickFalse`  is  the only  dependable  value of  type
`MagickBooleanType`; this fact is now reflected in the description
of the function `IsMagickCoreInstantiated'.

Technically,  this commit  alters  the  interface to  ImageMagick.
Before this commit, the description  implied code should do one of
the following:

  IsMagickCoreInstantiated() == MagickTrue
  IsMagickCoreInstantiated() != MagickTrue

Now, the description implies code should do one of the following:

  IsMagickCoreInstantiated() != MagickFalse
  IsMagickCoreInstantiated() == MagickFalse