genesis/terminus: Remove mutex from `MagickCoreGenesis()' and `MagickCoreTerminus()'
The  function   `MagickCoreGenesis'  has   hitherto  called   the  function
`InitializeMagickMutex', which must be called only in a synchronized manner
(i.e., it  must be  called by only  one thread  at a time,  and it  must be
called only once), essentially because the OpenMP implementation demands it.

Consequently, those  same requirements apply to  `MagickCoreGenesis' (i.e.,
it must be called by only one thread  at a time, and it must be called only
once);  clearly,  these  requirements  preclude  using  `MagickCoreGenesis'
multiple  times across  multiple threads,  and indeed  precludes using  the
mutex in question to synchronize reentrant calls of `MagickCoreGenesis'. In
other  words, the  current  synchronization  code that  uses  the mutex  is
incorrect, and the  idea that `MagickCoreGenesis' is some  kind of critical
section that must be synchronized is also wrong.

Therefore,   this   commit   removes    the   mutex   from   the   function
`MagickCoreGenesis'; really, the  user of this function  SHOULD be ensuring
that it is called only once to perform the desired initialization, and that
it is called in a way that does not cause reentrancy.

Of course, one can imagine that there already exists seemingly working code
where the function `MagickCoreGenesis' is  called multiple times, as in the
following:

  * A graphics  program starts  up by calling  `MagickCoreGenesis', because
    the program uses ImageMagick algorithms directly.

  * Next,  the  program loads  some  library  of graphics  algorithms,  and
    runs  that  library's  initialization  code, which  itself  also  calls
    `MagickCoreGenesis', because the library's programmers never considered
    the possibility that any other code would be using it.

Frankly, the library in this scenario  is broken; it should not be assuming
that it can  call `MagickCoreGenesis'. Fortunately, there  already exists a
way for the library to check whether  it should call it, namely by checking
the results of  a call to `IsMagickCoreInstantiated';  however, to decrease
the  problems of  a library  that does  not make  this check,  the function
`MagickCoreGenesis' errs on the side of  caution by doing the check itself.
It shouldn't be necessary to hold  programmers' hands in this way, but it's
perhaps better to  be safe than sorry  at this point, and  so this existing
check remains in place.

There is also  the possibility that `MagickCoreGenesis'  is called multiple
times in parallel:

  * A graphics  program does  not use ImageMagick  directly, but  creates 2
    threads, each of  which uses a different library of  algorithms that do
    use ImageMagick directly.

  * One thread  is started to load  and initialize one library,  whose code
    for initialization calls `MagickCoreGenesis'.

  * Simultaneously, another  thread is started  to load and  initialize the
    other library, whose code for initialization calls `MagickCoreGenesis'.

In this scenario, it's possible  that `MagickCoreGenesis' gets run not only
twice, but  concurrently, likely with  disastrous results. The only  way to
handle  this is  to treat  `MagickCoreGenesis' as  a critical  section that
requires mutual  exclusion, but  as already  noted, `InitializeMagickMutex'
would  have  to be  called  under  strict  conditions  before any  call  to
`MagickCoreGenesis',  so  that `MagickCoreGenesis'  could  use  a mutex  to
provide the necessary synchronization. Therefore, these strict conditions
might as well be applied to `MagickCoreGenesis' itself:

  If `IsMagickCoreInstantiated()  == MagickFalse', then one  (and only one)
  call to  `MagickCoreGenesis' must be  completed before any other  call to
  `MagickCoreGenesis' is made.

In  other  words, the  program  in  the  above  scenario must  be  declared
erroneous; the main  program must be rewritten  to call `MagickCoreGenesis'
in order  to prevent the  other threads  from doing anything  (please note,
this need  to rewrite is  already theoretically the case,  especially under
OpenMP; this commit  does not introduce a new requirement).  Here's how the
scenario should be resolved:

  * A graphics  program does  not use ImageMagick  directly, but  creates 2
    threads, each of  which uses a different library of  algorithms that do
    use ImageMagick directly.  The program must be aware of  this fact, and
    therefore call `MagickCoreGenesis' before starting those threads.

  * So, the program calls `MagickCoreGenesis' and allows it to complete.

  * Next, one thread  is started to load and initialize  one library, whose
    code for initialization calls  `MagickCoreGenesis'; that library's call
    to `MagickCoreGenesis' does nothing, because `MagickCoreGenesis' checks
    to see whether it has already done the necessary work, and sees that it
    has done so, thereby bailing out early with nary a side effect.

  * Simultaneously, another  thread is started  to load and  initialize the
    other library, whose code for initialization calls `MagickCoreGenesis';
    that library's call to `MagickCoreGenesis' also does nothing.

Shutting things down is more of a problem.

Unfortunately, having each of multiple components  of a program think it is
responsible for calling `MagickCoreTerminus'  cannot be handled gracefully.
Either those  components must be reconfigured  not to do this;  or, as with
`MagickCoreGenesis',  the main  program  must ensure  that  it alone  calls
`MagickCoreTerminus' before any other component  is allowed to do so, which
would allow any  further call of `MagickCoreTerminus' to become  a no-op as
with `MagickCoreGenesis'.  The main  problem with  this latter  approach is
that various components cannot be unloaded while other components are still
in use; again, no component that calls `MagickCoreTerminus' may be unloaded
until the main program calls `MagickCoreTerminus'.

So, you know, fix your end-user code.