memory: Update the description of `HeapOverflowSanityCheck()'
Firstly, the description was inverted in terms of the meaning of the
return value.

Secondly,  the   return  value  is  a   `MagickBooleanType';  the  only
dependenable value  of this type  is `MagickFalse', as is  evidenced by
the  way `MagickFalse'  is  generally used  throughout the  ImageMagick
source code, including with respect to the return value of calls to the
function `HeapOverflowSanityCheck'.

Consequently, the description of the function `HeapOverflowSanityCheck'
has been updated to reflect this fact.