memory: AcquireAlignedMemory: Move `AlignedExtent()` definition out of function body
The  idea   is  that   the  function-like  macro   `AlignedExtent'  is
being  defined for  use  in the  function `AcquireAlignedMemory',  and
so  the  macro  `AlignedExtent'  is  being  defined  lexically  inside
`AcquireAlignedMemory'; however, the nature of the C preprocessor, and
the fact  that the  definition of `AlignedExtent'  is not  indented to
match contents  of the  body of  `AcquireAlignedMemory' make  the code
look very awkward.

Therefore, this commit moves the definition of `AlignedExtent' outside
the body of `AcquireAlignedMemory'.