memory: Enhance the behavior of `HeapOverflowSanityCheck()'
Before, the function `HeapOverflowSanityCheck' would set `errno' to `ENOMEM'
if `count == 0'. However,  similar memory-related functions such as `calloc'
do not consider that to be an error;  after this commit, it is not longer an
error.

Also, this commit introduces a check as  to whether `quantum == 0', which is
treated  the same  as `count  == 0',  effectively also  testing whether  the
requested size (count*quantum) is zero.