gcc: Quash GCC's false positive warning for `-Wmaybe-uninitialized'
My tests indicate that this is  a problem only when using the `-Og'
optimization  level on  my system.  The nature  of this  warning is
currently  that  it could  produce  false  positives at  any  time,
depending on  the results  of optimizations and  the like,  so this
commit is more of a hack  to get around GCC's bad algorithm, rather
than a solution.

See here for more details:

  https://gcc.gnu.org/ml/gcc-patches/2019-11/msg02426.html

  Subject: Re: Move -Wmaybe-uninitialized to -Wextra
  Date: Tue, 26 Nov 2019 19:33:56 -0000
  Message-ID: <c50e4536644f4b4192afbe5995661ce4-mfwitten@gmail.com>