configure: Use `AR_FLAGS=cr' as the default
When GNU  `binutils' is configured  with `--enable-deterministic-archives',
the `D' modifier for the `ar'  tool becomes the default; unfortunately, the
maintainters of  `ar' thought it  would be cool to  tell as many  people as
possible about  their new  little feature,  yielding the  following warning
when the `u' flag  is used in conjunction with the  default `D' flag during
the linking of a static library.

  ar: `u' modifier ignored since `D' is the default (see `U')

Naturally,  other   extremely  long-lived  tools  such   as  `libtool'  and
`automake' (from  GNU, no  less) have  not been updated  to take  this into
account, and have long been using `cru' (that is, have long been using `u')
as an argument to `ar',  thereby provoking the aforementioned warning where
applicable.

The simplest solution is just to remove `u' from the invokation of `ar'; if
this causes  problems for anybody,  then they are probably  doing something
highly specialized, and can just invoke `configure' with their own options:

  ./configure AR_FLAGS=cru --prefix=...

You can read more about this here:

  https://www.mail-archive.com/bug-automake@gnu.org/msg04235.html
  Fri, 27 Mar 2015 08:49:35 -0700
  Pavel Raiskup