threshold: Embed `config/thresholds.xml' when `--enable-zero-configuration'
When ImageMagick  is configured with `--enable-zero-configuration',  it was
the case  that only a  very limited "file"  of threshold maps  was embedded
into ImageMagick for  use by the function `GetThresholdMap';  the maps were
so limited that it actually caused a failure during `make check' (the "2x2"
map was unrecognized by `-ordered-dither').

Now,  when `--enable-zero-configuration'  is  used, the  following file  is
embedded into ImageMagick:

  config/thresholds.xml

Of course,  the efficiency of this  could be improved by  massaging the XML
data (e.g., by  removing comments or unnecessary white space),  or by using
an entirely different  format for embedding.

This has only been implemented for platforms that support GNU's autotools.