coders/pnm: return status in WritePNMImage
Previously the code would always just return MagickTrue, even if it had
just inferred that the processing failed. This is required for more
elegant error handling throughout the function, such as checks on the
WriteBlob() calls, where the return value is currently being ignored.

Upon WriteBlob errors, we can then set the status to MagickFalse, and
break.