[OSS-Fuzz] Build fuzzers faster and resolve compiler warnings (#3064)
* use lld instead of GNU ld for linking the fuzzers

The fuzzers link a huge number of symbols (each fuzzer is ~43MB as of
this writing) and GNU ld is wildly slower at this. This change lowers
the build time for all fuzzers by 50% on a 22 core machine.

* == is not a valid comparison for char * in C

This worked anyway since the compiler unified the macro and the string
literal so it compared two identical addresses, but we should do the
comparison correctly.