Set color encoding in jxl encoder
libjxl requires the color encoding to be set before being able to encode
the pixels. If an ICC is available this patch sets the ICC profile on
the encoded image, otherwise it assumes sRGB.

In the future other image metadata like EXIF should be added here as
well.

Tested encoding a PNG with an ICC profile with `convert` and decoding it
back with `djxl`, colors look the same afterwards.