Update libjxl to use a .pc
Starting with libjxl v0.6 we can build static or shared libs when using
libjxl, so the check for `enable_static` is not needed anymore.

Also, it is possible to just use the .pc file from libjxl.

When encoding, not all the basic_info fields were set to valid values
when using memset(). This patch uses `JxlEncoderInitBasicInfo` to
initialize them instead of memset().