Remove fix for BeOS r3 which breaks gcc build with external toolchain
Hi all, 

This piece of code seems obsolete as it fixed something in BeOS R3 (which is from 1998). However it breaks the tests if the project is build with exteranl toolchain (not the one that comes from the Linux distribution). As LD_LIBRARY_PATH is already set (couple of lines above), $SED (sed) loads different glibc causing it to core. This leaves LD_LIBRARY_PATH *NOT* properly set and once the test are run (with make check) libmagic is unable to find external dependencies like bzip2, zlib etc. 

As I saw that $SED is used after as well it may potentially core at some other places (which however didn't happen in my build as it seems it didn't pass from that part of the code). This wouldn't happen of course if sed is completely statically build.

Thanks