Libxml2 fixes (#7784)
* libxml2: Remove HTTP and FTP support via libxml2

libxml2's nanohttp and nanoftp modules are about to be removed.

* libxml2: Stop calling xmlCleanupParser

libxml2 might be used by other code or libraries in the current process,
so it's not safe to call xmlCleanupParser. This function should only be
called right before exiting a process and after making sure that all
other threads using libxml2 are stopped.

* libxml2: Rework libxml2 SAX parser callbacks

The SAX handlers seem to be based upon a complete copy of an ancient
version of libxml2's SAX handlers. Many of the handlers don't require
modification and should use the original handlers from libxml2 to
receive updates and security fixes.

The easiest approach is to stop passing custom user data to the SAX
parser, meaning the SAX callbacks receive an xmlParserCtxtPtr. The
"info" structs are installed in parser->_private. This allows to use the
original SAX callbacks from libxml2 directly which are initialized with
a call to xmlSAXVersion.

Most of the removed handlers are related to DTDs and entities:

- internalSubset
- hasInternalSubset
- hasExternalSubset
- resolveEntity
- getEntity
- entityDecl
- notationDecl
- attributeDecl
- elementDecl
- unparsedEntityDecl
- getParameterEntity
- externalSubset

Start/end document:

- startDocument
- endDocument

Mostly meaningless:

- isStandalone
- setDocumentLocator

Some cleanup code is moved from SVGEndDocument to DestroySVGInfo.
SVGKeyValuePairs and SVGProcessStyleElement now take a SVGInfo instead
of a void pointer.

Some no-op handlers are removed and set to NULL which has the same
effect:

- comment (in msl.c)
- reference
- ignorableWhitespace
- processingInstruction

The cdataBlock handler is set to SVGCharacters, making CDATA section
actually work.