Fix slow processing for TIFFs that lie about their page numbers.
Previously, the code tried to query for the page number tag, which is documented as being
"The page number of the page from which this image was scanned".
https://www.loc.gov/preservation/digital/formats/content/tiff_tags.shtml

However, we've identified TIFFs in the wild that have the page number set as 0
for all pages.

This causes the current code to never break early, and causes the rendering to be slow.

This commit replaces the querying of TIFFTAG_PAGENUMBER with a counter.

It otherwise keeps the original behavior and API contract.